Put in a shader for stars
It's just an unlit shader that returns the color given, and I checked sebastian's code and project, and he used the same kind of shader, and when I downloaded the project, it looked the same, but in the video it looks really nice because of a bloom effect that isn't actually in his project. Movie magic perhaps?